refactor(rect): extract Rect to layout/rect.rs
